+++ R/cite_fude.R
@@ -14,61 +14,55 @@
 #' @export
 cite_fude <- function(data) {
 
-  if (is.list(data)) {
-    if ("fude" %in% names(data)) {
-      issue_year <- data$fude$issue_year
-      boundary_edit_year <- data$fude$boundary_edit_year
-    } else {
-      if ("polygon_uuid" %in% names(data)) {
-        issue_year <- data$issue_year
-        boundary_edit_year <- data$boundary_edit_year
-      } else {
-        issue_year <- unlist(lapply(data, function(df) {
-          if ("issue_year" %in% names(df)) {
-            return(df$issue_year)
-          } else {
-            stop("The input data must be Fude Polygon data.")
-          }
-        }))
-        boundary_edit_year <- unlist(lapply(data, function(df) {
-          return(df$boundary_edit_year)
-        }))
-      }
-    }
+  if (is.data.frame(data)) {
+
+    issue_year <- unique(data$issue_year)
+    boundary_edit_year <- unique(data$boundary_edit_year)
+
+  } else if (is.list(data)) {
+
+    issue_year <- unique(unlist(sapply(data, `[[`, "issue_year")))
+    boundary_edit_year <- unique(unlist(sapply(data, `[[`, "boundary_edit_year")))
+
+  }
+
+  if (is.null(issue_year) & is.null(boundary_edit_year)) {
+    stop("The input data must be Fude Polygon data.")
+  }
+
+  parts_ja <- c()
+  if (!is.null(issue_year)) {
+    parts_ja <- c(parts_ja, glue::glue("\u300C\u7B46\u30DD\u30EA\u30B4\u30F3\u30C7\u30FC\u30BF\uFF08{paste(sort(issue_year), collapse = '\uFF0C')}\u5E74\u5EA6\u516C\u958B\uFF09\u300D"))
+  }
+  if (!is.null(boundary_edit_year)) {
+    parts_ja <- c(parts_ja, glue::glue("\u300C\u8FB2\u696D\u96C6\u843D\u5883\u754C\u30C7\u30FC\u30BF\uFF08{paste(sort(boundary_edit_year), collapse = '\uFF0C')}\u5E74\u5EA6\uFF09\u300D"))
+  }
+  combined_parts_ja <- if (length(parts_ja) > 1) {
+    paste(parts_ja, collapse = "\u304A\u3088\u3073")
+  } else if (length(parts_ja) == 1) {
+    parts_ja
   } else {
-    if (is.data.frame(data)) {
-      if ("issue_year" %in% names(data)) {
-        issue_year <- data$issue_year
-        boundary_edit_year <- data$boundary_edit_year
-      } else {
-        stop("The input data must be Fude Polygon data.")
-      }
-    } else {
-      stop("The input data must be Fude Polygon data.")
-    }
+    ""
+  }
+
+  parts_en <- c()
+  if (!is.null(issue_year)) {
+    parts_en <- c(parts_en, glue::glue("'Fude Polygon Data (released in FY{paste(sort(issue_year), collapse = ', ')})'"))
+  }
+  if (!is.null(boundary_edit_year)) {
+    parts_en <- c(parts_en, glue::glue("'Agricultural Community Boundary Data (FY{paste(sort(boundary_edit_year), collapse = ', ')})'"))
+  }
+  combined_parts_en <- if (length(parts_en) > 1) {
+    paste(parts_en, collapse = " and ")
+  } else if (length(parts_en) == 1) {
+    parts_en
+  } else {
+    ""
   }
 
   x <- list(
-    ja = sprintf(
-      "\u8FB2\u6797\u6C34\u7523\u7701\u300C\u7B46\u30DD\u30EA\u30B4\u30F3\u30C7\u30FC\u30BF\uFF08%s\u5E74\u5EA6\u516C\u958B\uFF09%s\u300D\u3092\u52A0\u5DE5\u3057\u3066\u4F5C\u6210\u3002",
-      paste(sort(unique(issue_year)), collapse = "\uFF0C"),
-      if (!is.null(boundary_edit_year)) {
-        sprintf("\u300D\u304A\u3088\u3073\u300C\u8FB2\u696D\u96C6\u843D\u5883\u754C\u30C7\u30FC\u30BF\uFF08%s\u5E74\u5EA6\uFF09",
-                paste(sort(unique(issue_year)), collapse = "\uFF0C"))
-      } else {
-        ""
-      }
-    ),
-    en = sprintf(
-      "Created by processing the Ministry of Agriculture, Forestry and Fisheries, 'Fude Polygon Data (released in FY%s)'%s.",
-      paste(sort(unique(issue_year)), collapse = ", "),
-      if (!is.null(boundary_edit_year)) {
-        sprintf(" and 'Agricultural Community Boundary Data (FY%s)'",
-                paste(sort(unique(issue_year)), collapse = ", "))
-      } else {
-        ""
-      }
-    )
+    ja = glue::glue("\u8FB2\u6797\u6C34\u7523\u7701{combined_parts_ja}\u3092\u52A0\u5DE5\u3057\u3066\u4F5C\u6210\u3002"),
+    en = glue::glue("Created by processing the Ministry of Agriculture, Forestry and Fisheries{if (combined_parts_en != '') ', ' else ''}{combined_parts_en}.")
   )
 
   return(x)